Move all config functionality into separate module
