+++ lib/commands.js
@@ -1,32 +1,18 @@
 'use strict'
 
-// natives
-const path = require('path')
-
 // packages
 const moment = require('moment')
 const jsonfile = require('jsonfile')
-const Promise = require('bluebird')
 
 // ours
 const db = require('./db.js')
 const helpers = require('./utils/helpers.js')
 const spinner = require('./utils/spinner.js')
+const configManager = require('./utils/configManager.js')
 
 // constants
 const TODAY = moment().format('YYYY-MM-DD')
 const NOW = moment().format('HH:mm')
-const CONFIG_FILE = path.join(__dirname, 'config.json')
-
-// CONFIG has stuff that can be changed in runtime, so
-const configLoaded = new Promise((resolve, reject) => {
-  jsonfile.readFile(CONFIG_FILE, (err, obj) => {
-    if (err) {
-      reject(err)
-    }
-    resolve(obj)
-  })
-})
 
 // constants has real un-changeable stuff, they are read-only
 const constants = require('./constants.json')
@@ -56,7 +42,7 @@ const nextUndoneAction = (args, options, logger) => {
 
 const setStart = (args, options, logger) => {
   const start = args.start || NOW
-  configLoaded
+  configManager.configLoaded
     .then((config) => {
       const payload = {
         date: TODAY,
@@ -96,7 +82,7 @@ const setBreak = (args, options, logger) => {
 // report functionality for both single and batch reporting
 const report = (args, options, logger, date) => {
   date = date || TODAY
-  configLoaded
+  configManager.configLoaded
     .then((config) => {
       if (options && options.all) {
         db
@@ -130,7 +116,7 @@ const report = (args, options, logger, date) => {
 
 // set the configuration in the config file
 const setConfig = (args, options, logger) => {
-  configLoaded
+  configManager.configLoaded
     .then((config) => {
       if (options.day) {
         config.HOURS_IN_A_WORK_DAY = options.day
@@ -153,7 +139,7 @@ const setConfig = (args, options, logger) => {
           : 'default'
         spinner.succeed(`Default database path is set to ${dbPath}\n`)
       }
-      jsonfile.writeFileSync(CONFIG_FILE, config)
+      jsonfile.writeFileSync(configManager.CONFIG_FILE, config)
     })
     .catch(spinner.fail)
     .finally(() => process.exit(0))
