+++ lib/db.js
@@ -1,44 +1,14 @@
 'use strict'
 
 // native
-const path = require('path')
 const fs = require('mz/fs')
-
-// packages
-const osHomedir = require('os-homedir')
-const jsonfile = require('jsonfile')
 
 // ours
 // constants
-const constants = require('./constants.json')
 const helpers = require('./utils/helpers.js')
 const spinner = require('./utils/spinner.js')
-const CONFIG_FILE = path.join(__dirname, 'config.json')
-
-if (!fs.existsSync(CONFIG_FILE)) {
-  spinner.fail("Config file couldn't be found")
-  process.exit(1)
-}
-
-// determine path to database file
-const configPath = jsonfile.readFileSync(CONFIG_FILE).DB_FILE_MAIN
-const defaultPath = path.join(osHomedir(), constants.DB_FILE_MAIN)
-const dbTestFile = path.join(__dirname, '..', 'tests', constants.DB_FILE_FOR_TESTS)
-const isTestMode = process.env.MORO_TEST_MODE === 'true'
-
-const dbFile = isTestMode
-  ? dbTestFile
-  : configPath === '' ? defaultPath : configPath
-
-// log info about test mode
-const logTestMode = () => {
-  if (isTestMode) {
-    spinner.info('moro running in test mode, a temporary db will be used').start()
-  }
-}
-
-// use a temporary db if in test mode
-logTestMode()
+const configManager = require('./utils/configManager.js')
+const dbFile = configManager.getDBFilePath()
 
 const knex = require('knex')({
   dialect: 'sqlite3',
@@ -49,8 +19,7 @@ const knex = require('knex')({
 })
 
 const removeDatabase = () => {
-  logTestMode()
-
+  configManager.logTestMode()
   return fs.exists(dbFile)
     .then((exists) => {
       if (!exists) {
@@ -197,7 +166,6 @@ const getFullReport = (knex, CONFIG) => {
 }
 
 module.exports = {
-  dbTestFile,
   createTable,
   getDateReport,
   updateDatabase,
